extract NeonWalRecord apply logic
